pyproject.toml
[project]
name = "ai-street-market-agents-py"
version = "0.1.0"
description = "Python demo agents for the AI Street Market"
requires-python = ">=3.12"
dependencies = [
    "streetmarket @ git+https://github.com/org-moredevs-ai/ai-street-market.git#subdirectory=libs",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = ["agents*"]
